pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "lumni"
version = "2.0.0"
description = "Unified Student API Gateway with smart fallback and limit monitoring"
readme = "README.md"
license = "MIT"
authors = ["Lumni Team"]
keywords = ["api-gateway", "ai", "student", "fallback", "limit-monitoring"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
package-mode = false

[tool.poetry.scripts]
lumni = "app.cli.main:app"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = ">=0.104.0"
uvicorn = {extras = ["standard"], version = ">=0.24.0"}
litellm = ">=1.50.0"
portkey-ai = ">=1.0.0"
sqlalchemy = ">=2.0.0"
alembic = ">=1.12.0"
psycopg2-binary = ">=2.9.0"
redis = ">=5.0.0"
pydantic = ">=2.5.0"
pydantic-settings = ">=2.1.0"
python-dotenv = ">=1.0.0"
httpx = ">=0.25.0"
pyyaml = ">=6.0.0"
tenacity = ">=8.2.0"
slowapi = ">=0.1.9"
prometheus-client = ">=0.19.0"
typer = {extras = ["all"], version = ">=0.9.0"}
rich = ">=13.0.0"
prompt-toolkit = ">=3.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.4.0"
pytest-asyncio = ">=0.21.0"
pytest-cov = ">=4.1.0"
black = ">=23.11.0"
ruff = ">=0.1.6"
mypy = ">=1.7.0"
nbformat = "^5.10.4"
ipykernel = "^7.1.0"
jupyterlab = "^4.5.0"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
